refactor(playground): split execution subscriber module
